Refactor input handling across multiple locations to replace line input with character input, enhancing user interaction consistency. Update related methods in TerminalEmulator for improved clarity and functionality. These changes aim to streamline user experience and maintain code clarity.

+++ Scripts/Locations/HealerLocation.cs
@@ -18,7 +18,7 @@ public partial class HealerLocation : BaseLocation
     {
         base._Ready();
         Name = "HealerLocation";
-        LocationId = (int)GameLocation.Healer;
+        LocationId = GameLocation.Healer;
     }
     
     protected new void ProcessPlayerInput(string input)
